SP-3: Change signature of PostService#getPost and implement this method
- add intermediate PostDataWrapper dto

+++ src/main/java/org/practicum/yandex/repository/PostRepository.java
@@ -9,7 +9,7 @@ import java.util.Set;
 public interface PostRepository extends SqlRepository<BigInteger, PostModel> {
     void saveTags(final BigInteger postId, final List<BigInteger> tagIds);
     List<BigInteger> getOrInsertTags(final List<String> tagNames);
-    Page<PostModel> findPaged(final int page, final int size, String query, Set<String> tags);
+    Page<PostModel> findPaged(final long page, final long size, String query, Set<String> tags);
     void removePostTags(final BigInteger postId);
     BigInteger incrementLikes(final BigInteger postId);
 }
